+++ bash.exe.stackdump
@@ -1,0 +1,29 @@
+Stack trace:
+Frame         Function      Args
+0007FFFF9AF0  00021005FE8E (000210285F68, 00021026AB6E, 000000000000, 0007FFFF89F0) msys-2.0.dll+0x1FE8E
+0007FFFF9AF0  0002100467F9 (000000000000, 000000000000, 000000000000, 0007FFFF9DC8) msys-2.0.dll+0x67F9
+0007FFFF9AF0  000210046832 (000210286019, 0007FFFF99A8, 000000000000, 000000000000) msys-2.0.dll+0x6832
+0007FFFF9AF0  000210068CF6 (000000000000, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x28CF6
+0007FFFF9AF0  000210068E24 (0007FFFF9B00, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x28E24
+0007FFFF9DD0  00021006A225 (0007FFFF9B00, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x2A225
+End of stack trace
+Loaded modules:
+000100400000 bash.exe
+7FFC69390000 ntdll.dll
+7FFC67F00000 KERNEL32.DLL
+7FFC66690000 KERNELBASE.dll
+7FFC68F80000 USER32.dll
+7FFC66BB0000 win32u.dll
+7FFC68EE0000 GDI32.dll
+7FFC66DD0000 gdi32full.dll
+7FFC664A0000 msvcp_win.dll
+000210040000 msys-2.0.dll
+7FFC66570000 ucrtbase.dll
+7FFC68350000 advapi32.dll
+7FFC681D0000 msvcrt.dll
+7FFC67B00000 sechost.dll
+7FFC66540000 bcrypt.dll
+7FFC68420000 RPCRT4.dll
+7FFC65BA0000 CRYPTBASE.DLL
+7FFC66D50000 bcryptPrimitives.dll
+7FFC692F0000 IMM32.DLL
